+++ changelog.md
@@ -1,3 +1,6 @@
+##Kotlin for Forge 1.12.2
+- Added support for using Mod.EventBusSubscriber as a file annotation. See AutoKotlinEventBusSubscriber.kt KDoc for example.
+
 ##Kotlin for Forge 1.12.1
 - Updated to Kotlin 1.5.10, Updated to coroutines 1.5.0, Updated to serialization 1.2.1
 - Added a missing serialization library that went missing in 1.12.0